Preparing the structure to make it work with different versions of Django
