+++ .travis.yml
@@ -7,6 +7,7 @@ env:
   - TOXENV=django14
   - TOXENV=django15
   - TOXENV=django16
+  - TOXENV=django17
 
 script:
   - tox
