+++ deep_collector/compat/serializers/django_1_6.py
