+++ deep_collector/compat/serializers/django_1_7.py
@@ -1,0 +1,5 @@
+
+from django.core.serializers.json import Serializer
+
+
+MultiModelInheritanceSerializer = Serializer
