+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = django14,django15,django16,django17,django18
+envlist = django14,django15,django16,django17
 
 [testenv]
 commands = python runtests.py
@@ -9,11 +9,13 @@ deps =
 [testenv:django14]
 deps =
     django>=1.4,<1.4.99
+    django-discover-runner
     {[testenv]deps}
 
 [testenv:django15]
 deps =
     django>=1.5,<1.5.99
+    django-discover-runner
     {[testenv]deps}
 
 [testenv:django16]
@@ -24,9 +26,4 @@ deps =
 [testenv:django17]
 deps =
     django>=1.7,<1.7.99
-    {[testenv]deps}
-
-[testenv:django18]
-deps = 
-    django>=1.8,<1.8.99
     {[testenv]deps}
